Learnli/requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
billiard==4.2.0
boto3==1.34.154
botocore==1.34.154
cachetools==5.4.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
distro==1.9.0
dj-database-url==2.3.0
Django==5.1.1
django-chartjs==2.3.0
django-ckeditor==6.7.1
django-crispy-forms==2.3
django-extensions==3.2.3
django-formtools==2.5.1
django-js-asset==2.2.0
django-otp==1.5.2
django-phonenumber-field==8.0.0
django-two-factor-auth==1.17.0
djangoflutterwave==0.3.0
djangorestframework==3.15.1
et-xmlfile==1.1.0
fonttools==4.46.0
frozenlist==1.4.1
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-core==2.4.1
google-cloud-speech==2.26.1
google-cloud-translate==3.15.4
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.65.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.7
jiter==0.7.1
jmespath==1.0.1
kiwisolver==1.4.5
kombu==5.4.0
matplotlib==3.8.2
multidict==6.1.0
numpy==1.26.2
openai==1.54.5
openpyxl==3.1.2
packaging==23.2
phonenumbers==8.13.44
Pillow==9.4.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.2
psycopg2-binary==2.9.10
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyparsing==3.1.1
pypng==0.20220715.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
qrcode==7.4.2
redis==5.1.0
reportlab==3.6.12
requests==2.32.3
rsa==4.9
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.3
stripe==10.5.0
tqdm==4.67.0
twilio==9.3.2
typing_extensions==4.12.2
tzdata==2022.7
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
yarl==1.13.1
